Wire up cobra commands. Move exercises out cmd/, int app/
