+++ Makefile
@@ -1,8 +1,8 @@
 test:
-	(cd go/cmd/exercises && go tool gotestsum)
+	(cd go && go tool gotestsum)
 
 lint:
-	(cd go/cmd/exercises && go tool golangci-lint run)
+	(cd go && go tool golangci-lint run)
 
 format:
-	(cd go/cmd/exercises && go tool golangci-lint fmt)
+	(cd go && go tool golangci-lint fmt)
