+++ go/app/exercises/01-two-sums/two_sums_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/iainjp/blind75/cmd/utils"
+	"github.com/iainjp/blind75/utils"
 )
 
 func TestTwoSums(t *testing.T) {
