+++ go/app/exercises/02-time-to-buy-stocks/README.md
