+++ go/app/exercises/04-product-of-array-except-self/README.md
