+++ go/app/exercises/04-product-of-array-except-self/product_except_self.go
