+++ go/app/exercises/04-product-of-array-except-self/product_except_self_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/iainjp/blind75/cmd/utils"
+	"github.com/iainjp/blind75/utils"
 )
 
 func TestProductExceptSelf(t *testing.T) {
