+++ go/app/exercises/05-maximum-subarray/README.md
