+++ go/app/exercises/05-maximum-subarray/maximum_subarray.go
