+++ go/app/exercises/05-maximum-subarray/maximum_subarray_test.go
@@ -4,23 +4,23 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/iainjp/blind75/cmd/utils"
+	"github.com/iainjp/blind75/utils"
 )
 
-func TestMinRotatedSortedArray(t *testing.T) {
+func TestMaximumSubarray(t *testing.T) {
 	var testCases = []struct {
 		input []int
 		want  int
 	}{
-		{[]int{3, 4, 5, 1, 2}, 1},
-		{[]int{4, 5, 6, 7, 0, 1, 2}, 0},
-		{[]int{11, 13, 15, 17}, 11},
+		{[]int{-2, 1, -3, 4, -1, 2, 1, -5, 4}, 6},
+		{[]int{1}, 1},
+		{[]int{5, 4, -1, 7, 8}, 23},
 	}
 
 	for testNum, test := range testCases {
 		testName := fmt.Sprintf("Case [%v]", testNum)
 		t.Run(testName, func(t *testing.T) {
-			got := findMin(test.input)
+			got := maxSubArray(test.input)
 			utils.CheckEqual(got, test.want, t)
 		})
 	}
