+++ go/app/exercises/06-maximum-product-subarray/README.md
