+++ go/app/exercises/06-maximum-product-subarray/maximum_product_subarray.go
