+++ go/app/exercises/06-maximum-product-subarray/maximum_product_subarray_test.go
@@ -4,23 +4,22 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/iainjp/blind75/cmd/utils"
+	"github.com/iainjp/blind75/utils"
 )
 
-func TestContainsDuplicates(t *testing.T) {
+func TestMaximumProductSubarray(t *testing.T) {
 	var testCases = []struct {
 		input []int
-		want  bool
+		want  int
 	}{
-		{[]int{1, 2, 3, 1}, true},
-		{[]int{1, 2, 3, 4}, false},
-		{[]int{1, 1, 1, 3, 3, 4, 3, 2, 4, 2}, true},
+		{[]int{2, 3, -2, 4}, 6},
+		{[]int{-2, 0, -1}, 0},
 	}
 
 	for testNum, test := range testCases {
 		testName := fmt.Sprintf("Case [%v]", testNum)
 		t.Run(testName, func(t *testing.T) {
-			got := containsDuplicate(test.input)
+			got := maxProduct(test.input)
 			utils.CheckEqual(got, test.want, t)
 		})
 	}
