+++ go/app/exercises/50-length-of-longest-substring/longest_substring_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/iainjp/blind75/cmd/utils"
+	"github.com/iainjp/blind75/utils"
 )
 
 func TestLongestSubstring(t *testing.T) {
