+++ go/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/iainjp/blind75/cmd"
+
+func main() {
+	cmd.Execute()
+}
